Build A Simple Product.
Products will be what the application mostly surrounds itself with.
Create the overarching product model using the model scaffold provided
by Rails.

Some folks take issue with the scaffolding part of Rails but, really it
does a lot of nice defaults for a developer. Not convinced that creating
these files 'by hand' makes you a worse or better programmer.