Inertia JS For Root.
Boot the root Svelte app at the root URL of the Rails application. This
alters the return from the Product#index to utilize Inertia as well as
booting the actual root application once and only once the DOM Content
has loaded. Silly mistake (event listener) on my side.

This also includes a very bare bones, simple system test. Being an 'SPA'
of sorts, the application is going to rely heavily on system tests + JS
testing for the Svelte components.