Implement Show URL For Products.
To be able to one click buy a product, there needs to be some kind of
product display page or 'PDP'. Implement this in the products controller
and pass the relevant information surrounding the product to Inertia.
This page doesn't exist yet so the product system test around this route
will fail (not committed) but, the integration test for the controller
does pass.

This also includes a few bits of cleanup to bring the product
integration test inline with minitest spec versus the assertion
matchers.