Create A Svelte Layout.
Create a layout that will be persisted throughout the entire SPA. At the
moment it's pretty simple, just some links and a product container.
However, it will turn into the epicenter for the application as it
progresses.

This also includes a change to the root layout that tests the presence
of the layout elements versus a default string.