Render A Simple Product Page.
Get product detail pages rendering properly when navigating to them.
This includes a small system test and an extremely un-styled PDP page.
Tailwind CSS to come to take care of that.

Honestly, I am not convinced that the way props are accessed is correct.
There isn't too much information surrounding this in the Inertia docs
but, for now it 'works'.